rename to wfrp
