+++ bestiary/wfrp/gen_statblocks.py
