+++ bestiary/wfrp/scraper.py
